Modified generic 500 error responses from routes to include the actual error that caused the message
