--- src/routes/error_body.rs
@@ -1,15 +1,0 @@
-//! Defines a struct to use for error messages return by the rest mapping
-
-use serde::{Deserialize, Serialize};
-
-///Struct to use for returning error responses from REST endpoints
-///
-/// `title` is a brief summary of the error
-/// `status` is the http status code
-/// `detail` is a more detailed explanation of the error
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-pub struct ErrorBody {
-    pub title: String,
-    pub status: u16,
-    pub detail: String,
-}
