+++ src/routes/pipeline.rs
@@ -5,7 +5,7 @@
 
 use crate::db;
 use crate::models::pipeline::{NewPipeline, PipelineChangeset, PipelineData, PipelineQuery};
-use crate::routes::error_body::ErrorBody;
+use crate::routes::error_handling::{ ErrorBody, default_500 };
 use actix_web::{error::BlockingError, web, HttpRequest, HttpResponse, Responder};
 use log::error;
 use serde_json::json;
@@ -63,11 +63,7 @@ async fn find_by_id(
                 detail: "No pipeline found with the specified ID".to_string(),
             }),
             // For other errors, return a 500
-            _ => HttpResponse::InternalServerError().json(ErrorBody {
-                title: "Server error".to_string(),
-                status: 500,
-                detail: "Error while attempting to retrieve requested pipeline from DB".to_string(),
-            }),
+            _ => default_500(&e),
         }
     })?;
 
@@ -116,11 +112,7 @@ async fn find(
     .map_err(|e| {
         error!("{}", e);
         // If there is an error, return a 500
-        HttpResponse::InternalServerError().json(ErrorBody {
-            title: "Server error".to_string(),
-            status: 500,
-            detail: "Error while attempting to retrieve requested pipeline(s) from DB".to_string(),
-        })
+        default_500(&e)
     })?;
 
     Ok(res)
@@ -157,11 +149,7 @@ async fn create(
     .map_err(|e| {
         error!("{}", e);
         // If there is an error, return a 500
-        HttpResponse::InternalServerError().json(ErrorBody {
-            title: "Server error".to_string(),
-            status: 500,
-            detail: "Error while attempting to insert new pipeline".to_string(),
-        })
+        default_500(&e)
     })?;
     Ok(res)
 }
@@ -212,11 +200,7 @@ async fn update(
     .map_err(|e| {
         error!("{}", e);
         // If there is an error, return a 500
-        HttpResponse::InternalServerError().json(ErrorBody {
-            title: "Server error".to_string(),
-            status: 500,
-            detail: "Error while attempting to update pipeline".to_string(),
-        })
+        default_500(&e)
     })?;
 
     Ok(res)
@@ -289,11 +273,7 @@ async fn delete_by_id(req: HttpRequest, pool: web::Data<db::DbPool>) -> impl Res
                 detail: "Cannot delete a pipeline if there is a template mapped to it".to_string(),
             }),
             // For other errors, return a 500
-            _ => HttpResponse::InternalServerError().json(ErrorBody {
-                title: "Server error".to_string(),
-                status: 500,
-                detail: "Error while attempting to delete requested pipeline from DB".to_string(),
-            }),
+            _ => default_500(&e),
         }
     })
 }
@@ -532,10 +512,6 @@ mod tests {
 
         assert_eq!(error_body.title, "Server error");
         assert_eq!(error_body.status, 500);
-        assert_eq!(
-            error_body.detail,
-            "Error while attempting to insert new pipeline"
-        );
     }
 
     #[actix_rt::test]
@@ -628,10 +604,6 @@ mod tests {
 
         assert_eq!(error_body.title, "Server error");
         assert_eq!(error_body.status, 500);
-        assert_eq!(
-            error_body.detail,
-            "Error while attempting to update pipeline"
-        );
     }
 
     #[actix_rt::test]
