+++ src/routes/software.rs
@@ -5,7 +5,7 @@
 
 use crate::db;
 use crate::models::software::{NewSoftware, SoftwareChangeset, SoftwareData, SoftwareQuery};
-use crate::routes::error_body::ErrorBody;
+use crate::routes::error_handling::{ ErrorBody, default_500 };
 use crate::util::git_repos;
 use actix_web::{error::BlockingError, web, HttpRequest, HttpResponse};
 use log::error;
@@ -63,11 +63,7 @@ async fn find_by_id(
                 detail: "No software found with the specified ID".to_string(),
             }),
             // For other errors, return a 500
-            _ => HttpResponse::InternalServerError().json(ErrorBody {
-                title: "Server error".to_string(),
-                status: 500,
-                detail: "Error while attempting to retrieve requested software from DB".to_string(),
-            }),
+            _ => default_500(&e),
         }
     })?;
 
@@ -116,11 +112,7 @@ async fn find(
     .map_err(|e| {
         error!("{}", e);
         // If there is an error, return a 500
-        HttpResponse::InternalServerError().json(ErrorBody {
-            title: "Server error".to_string(),
-            status: 500,
-            detail: "Error while attempting to retrieve requested software(s) from DB".to_string(),
-        })
+        default_500(&e)
     })?;
 
     Ok(res)
@@ -185,11 +177,7 @@ async fn create(
     .map_err(|e| {
         error!("{}", e);
         // If there is an error, return a 500
-        HttpResponse::InternalServerError().json(ErrorBody {
-            title: "Server error".to_string(),
-            status: 500,
-            detail: "Error while attempting to insert new software".to_string(),
-        })
+        default_500(&e)
     })?;
     Ok(res)
 }
@@ -240,11 +228,7 @@ async fn update(
     .map_err(|e| {
         error!("{}", e);
         // If there is an error, return a 500
-        HttpResponse::InternalServerError().json(ErrorBody {
-            title: "Server error".to_string(),
-            status: 500,
-            detail: "Error while attempting to update software".to_string(),
-        })
+        default_500(&e)
     })?;
 
     Ok(res)
@@ -271,7 +255,7 @@ pub fn init_routes(cfg: &mut web::ServiceConfig) {
 mod tests {
 
     use super::*;
-    use crate::routes::error_body::ErrorBody;
+    use crate::routes::error_handling::ErrorBody;
     use crate::unit_test_util::*;
     use actix_web::{http, test, App};
     use diesel::PgConnection;
@@ -473,10 +457,6 @@ mod tests {
 
         assert_eq!(error_body.title, "Server error");
         assert_eq!(error_body.status, 500);
-        assert_eq!(
-            error_body.detail,
-            "Error while attempting to insert new software"
-        );
     }
 
     #[actix_rt::test]
@@ -604,9 +584,5 @@ mod tests {
 
         assert_eq!(error_body.title, "Server error");
         assert_eq!(error_body.status, 500);
-        assert_eq!(
-            error_body.detail,
-            "Error while attempting to update software"
-        );
     }
 }
