+++ src/routes/template_report.rs
@@ -7,7 +7,7 @@ use crate::db;
 use crate::models::template_report::{
     DeleteError, NewTemplateReport, TemplateReportData, TemplateReportQuery,
 };
-use crate::routes::error_body::ErrorBody;
+use crate::routes::error_handling::{ ErrorBody, default_500 };
 use actix_web::{error::BlockingError, web, HttpRequest, HttpResponse, Responder};
 use log::error;
 use serde::{Deserialize, Serialize};
@@ -96,12 +96,7 @@ async fn find_by_id(req: HttpRequest, pool: web::Data<db::DbPool>) -> impl Respo
                 detail: "No template_report mapping found with the specified ID".to_string(),
             }),
             // For other errors, return a 500
-            _ => HttpResponse::InternalServerError().json(ErrorBody {
-                title: "Server error".to_string(),
-                status: 500,
-                detail: "Error while attempting to retrieve requested template_report from DB"
-                    .to_string(),
-            }),
+            _ => default_500(&e),
         }
     })
 }
@@ -169,11 +164,7 @@ async fn find(
     .map_err(|e| {
         error!("{}", e);
         // For any errors, return a 500
-        HttpResponse::InternalServerError().json(ErrorBody {
-            title: "Server error".to_string(),
-            status: 500,
-            detail: "Error while attempting to retrieve requested mapping(s) from DB".to_string(),
-        })
+        default_500(&e)
     })
 }
 
@@ -251,11 +242,7 @@ async fn create(
     .map_err(|e| {
         error!("{}", e);
         // For any errors, return a 500
-        HttpResponse::InternalServerError().json(ErrorBody {
-            title: "Server error".to_string(),
-            status: 500,
-            detail: "Error while attempting to insert new template report mapping".to_string(),
-        })
+        default_500(&e)
     })
 }
 
@@ -341,11 +328,7 @@ async fn delete_by_id(req: HttpRequest, pool: web::Data<db::DbPool>) -> impl Res
                     detail: "Cannot delete a template_report mapping if the associated template has non-failed run that has a non-failed run_report from the associated report".to_string(),
                 }),
                 // For other errors, return a 500
-                _ => HttpResponse::InternalServerError().json(ErrorBody {
-                    title: "Server error".to_string(),
-                    status: 500,
-                    detail: "Error while attempting to delete requested template_report mapping from DB".to_string(),
-                }),
+                _ => default_500(&e),
             }
         })
 }
